(fix|COS-2000): Fixed updating of arr after renewal likelihood change
